Separete subquery structure related optimizer rules into 3 files
