+++ arangod/Aql/Optimizer/Rule/InlineSubqueries.cpp
@@ -1,0 +1,250 @@
+////////////////////////////////////////////////////////////////////////////////
+/// DISCLAIMER
+///
+/// Copyright 2014-2026 ArangoDB GmbH, Cologne, Germany
+/// Copyright 2004-2014 triAGENS GmbH, Cologne, Germany
+///
+/// Licensed under the Business Source License 1.1 (the "License");
+/// you may not use this file except in compliance with the License.
+/// You may obtain a copy of the License at
+///
+///     https://github.com/arangodb/arangodb/blob/devel/LICENSE
+///
+/// Unless required by applicable law or agreed to in writing, software
+/// distributed under the License is distributed on an "AS IS" BASIS,
+/// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+/// See the License for the specific language governing permissions and
+/// limitations under the License.
+///
+/// Copyright holder is ArangoDB GmbH, Cologne, Germany
+///
+/// @author Max Neunhoeffer
+/// @author Jan Steemann
+////////////////////////////////////////////////////////////////////////////////
+
+#include "InlineSubqueries.h"
+
+#include "Aql/Ast.h"
+#include "Aql/ExecutionNode/CalculationNode.h"
+#include "Aql/ExecutionNode/CollectNode.h"
+#include "Aql/ExecutionNode/EnumerateListNode.h"
+#include "Aql/ExecutionNode/ExecutionNode.h"
+#include "Aql/ExecutionNode/ReturnNode.h"
+#include "Aql/ExecutionNode/SubqueryNode.h"
+#include "Aql/ExecutionNode/WindowNode.h"
+#include "Aql/ExecutionPlan.h"
+#include "Aql/Expression.h"
+#include "Aql/Optimizer.h"
+#include "Aql/Optimizer/Utils/VariableReplacer.h"
+#include "Aql/Variable.h"
+#include "Aql/types.h"
+#include "Containers/SmallVector.h"
+
+namespace arangodb::aql {
+using EN = ExecutionNode;
+
+/// @brief pulls out simple subqueries and merges them with the level above
+///
+/// For example, if we have the input query
+///
+/// FOR x IN (
+///     FOR y IN collection FILTER y.value >= 5 RETURN y.test
+///   )
+///   RETURN x.a
+///
+/// then this rule will transform it into:
+///
+/// FOR tmp IN collection
+///   FILTER tmp.value >= 5
+///   LET x = tmp.test
+///   RETURN x.a
+void inlineSubqueriesRule(Optimizer* opt, std::unique_ptr<ExecutionPlan> plan,
+                          OptimizerRule const& rule) {
+  containers::SmallVector<ExecutionNode*, 8> nodes;
+  plan->findNodesOfType(nodes, EN::SUBQUERY, true);
+
+  if (nodes.empty()) {
+    opt->addPlan(std::move(plan), rule, false);
+    return;
+  }
+
+  bool modified = false;
+  std::vector<ExecutionNode*> subNodes;
+
+  for (auto const& n : nodes) {
+    auto subqueryNode = ExecutionNode::castTo<SubqueryNode*>(n);
+
+    if (subqueryNode->isModificationNode()) {
+      continue;
+    }
+
+    if (!subqueryNode->isDeterministic()) {
+      continue;
+    }
+
+    bool eligible = true;
+    bool containsLimitOrSort = false;
+    auto current = subqueryNode->getSubquery();
+    TRI_ASSERT(current != nullptr);
+
+    while (current != nullptr) {
+      if (current->getType() == EN::WINDOW && subqueryNode->isInInnerLoop()) {
+        eligible = false;
+        break;
+      } else if (current->getType() == EN::COLLECT) {
+        if (subqueryNode->isInInnerLoop()) {
+          eligible = false;
+          break;
+        }
+        if (ExecutionNode::castTo<CollectNode const*>(current)
+                ->hasOutVariable()) {
+          eligible = false;
+          break;
+        }
+      } else if (current->getType() == EN::LIMIT ||
+                 current->getType() == EN::SORT) {
+        containsLimitOrSort = true;
+      }
+      current = current->getFirstDependency();
+    }
+
+    if (!eligible) {
+      continue;
+    }
+
+    Variable const* out = subqueryNode->outVariable();
+    TRI_ASSERT(out != nullptr);
+    VarSet subqueryVars;
+    subqueryVars.emplace(out);
+
+    std::vector<ExecutionNode*> aliasNodesToRemoveLater;
+
+    VarSet varsUsed;
+
+    current = n->getFirstParent();
+    while (current->hasParent()) {
+      if (current->getType() == EN::ENUMERATE_LIST) {
+        if (current->isInInnerLoop() && containsLimitOrSort) {
+          current = nullptr;
+          break;
+        }
+
+        auto listNode = ExecutionNode::castTo<EnumerateListNode*>(current);
+        if (listNode->getMode() == EnumerateListNode::kEnumerateObject) {
+          current = nullptr;
+          break;
+        }
+        if (subqueryVars.find(listNode->inVariable()) != subqueryVars.end()) {
+          bool mustAbort = false;
+          for (auto const& itSub : subqueryVars) {
+            if (listNode->isVarUsedLater(itSub)) {
+              current = nullptr;
+              mustAbort = true;
+              break;
+            }
+          }
+          if (mustAbort) {
+            break;
+          }
+
+          for (auto const& toRemove : aliasNodesToRemoveLater) {
+            plan->unlinkNode(toRemove, false);
+          }
+
+          subNodes.clear();
+          subNodes.reserve(4);
+          subqueryNode->getSubquery()->getDependencyChain(subNodes, true);
+          TRI_ASSERT(!subNodes.empty());
+          auto returnNode = ExecutionNode::castTo<ReturnNode*>(subNodes[0]);
+          TRI_ASSERT(returnNode->getType() == EN::RETURN);
+
+          modified = true;
+          auto queryVariables = plan->getAst()->variables();
+          auto previous = n->getFirstDependency();
+          auto insert = n->getFirstParent();
+          TRI_ASSERT(insert != nullptr);
+
+          plan->unlinkNode(n, false);
+
+          for (auto& it : subNodes) {
+            plan->unlinkNode(it, true);
+
+            if (it->getType() == EN::SINGLETON) {
+              break;
+            }
+
+            if (it != returnNode) {
+              insert->removeDependencies();
+              TRI_ASSERT(it != nullptr);
+              insert->addDependency(it);
+              insert = it;
+
+              for (auto const& variable : it->getVariablesSetHere()) {
+                queryVariables->renameVariable(variable->id);
+              }
+            }
+          }
+
+          if (previous != nullptr) {
+            insert->addDependency(previous);
+          }
+
+          plan->unlinkNode(listNode, false);
+
+          queryVariables->renameVariable(returnNode->inVariable()->id,
+                                         listNode->outVariable()[0]->name);
+
+          std::unordered_map<VariableId, Variable const*> replacements;
+          replacements.try_emplace(listNode->outVariable()[0]->id,
+                                   returnNode->inVariable());
+          VariableReplacer finder(replacements);
+          plan->root()->walk(finder);
+
+          plan->clearVarUsageComputed();
+          plan->findVarUsage();
+
+          current = nullptr;
+        }
+      } else if (current->getType() == EN::CALCULATION) {
+        auto rootNode = ExecutionNode::castTo<CalculationNode*>(current)
+                            ->expression()
+                            ->node();
+        if (rootNode->type == NODE_TYPE_REFERENCE) {
+          if (subqueryVars.find(static_cast<Variable const*>(
+                  rootNode->getData())) != subqueryVars.end()) {
+            subqueryVars.emplace(
+                ExecutionNode::castTo<CalculationNode*>(current)
+                    ->outVariable());
+            aliasNodesToRemoveLater.emplace_back(current);
+            current = current->getFirstParent();
+
+            continue;
+          }
+        }
+      }
+
+      if (current == nullptr) {
+        break;
+      }
+
+      varsUsed.clear();
+      current->getVariablesUsedHere(varsUsed);
+
+      bool mustAbort = false;
+      for (auto const& itSub : subqueryVars) {
+        if (varsUsed.find(itSub) != varsUsed.end()) {
+          mustAbort = true;
+          break;
+        }
+      }
+      if (mustAbort) {
+        break;
+      }
+
+      current = current->getFirstParent();
+    }
+  }
+
+  opt->addPlan(std::move(plan), rule, modified);
+}
+}  // namespace arangodb::aql
