+++ arangod/Aql/Optimizer/Rule/OptimizeSubqueries.cpp
@@ -1,0 +1,221 @@
+////////////////////////////////////////////////////////////////////////////////
+/// DISCLAIMER
+///
+/// Copyright 2014-2026 ArangoDB GmbH, Cologne, Germany
+/// Copyright 2004-2014 triAGENS GmbH, Cologne, Germany
+///
+/// Licensed under the Business Source License 1.1 (the "License");
+/// you may not use this file except in compliance with the License.
+/// You may obtain a copy of the License at
+///
+///     https://github.com/arangodb/arangodb/blob/devel/LICENSE
+///
+/// Unless required by applicable law or agreed to in writing, software
+/// distributed under the License is distributed on an "AS IS" BASIS,
+/// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+/// See the License for the specific language governing permissions and
+/// limitations under the License.
+///
+/// Copyright holder is ArangoDB GmbH, Cologne, Germany
+///
+/// @author Max Neunhoeffer
+/// @author Jan Steemann
+////////////////////////////////////////////////////////////////////////////////
+
+#include "OptimizeSubqueries.h"
+
+#include "Aql/Ast.h"
+#include "Aql/ExecutionNode/CalculationNode.h"
+#include "Aql/ExecutionNode/ExecutionNode.h"
+#include "Aql/ExecutionNode/LimitNode.h"
+#include "Aql/ExecutionNode/ReturnNode.h"
+#include "Aql/ExecutionNode/SubqueryNode.h"
+#include "Aql/ExecutionPlan.h"
+#include "Aql/Expression.h"
+#include "Aql/Optimizer.h"
+#include "Aql/TypedAstNodes.h"
+#include "Aql/Variable.h"
+#include "Aql/types.h"
+#include "Containers/SmallVector.h"
+
+#include <tuple>
+#include <unordered_map>
+#include <unordered_set>
+
+namespace arangodb::aql {
+using EN = ExecutionNode;
+
+void optimizeSubqueriesRule(Optimizer* opt, std::unique_ptr<ExecutionPlan> plan,
+                            OptimizerRule const& rule) {
+  bool modified = false;
+
+  containers::SmallVector<ExecutionNode*, 8> nodes;
+  plan->findNodesOfType(nodes, EN::CALCULATION, true);
+
+  std::unordered_map<
+      ExecutionNode*,
+      std::tuple<int64_t, std::unordered_set<ExecutionNode const*>, bool>>
+      subqueryAttributes;
+
+  for (auto const& n : nodes) {
+    auto cn = ExecutionNode::castTo<CalculationNode*>(n);
+    AstNode const* root = cn->expression()->node();
+    if (root == nullptr) {
+      continue;
+    }
+
+    auto visitor = [&subqueryAttributes, &plan,
+                    n](AstNode const* node) -> bool {
+      std::pair<ExecutionNode*, int64_t> found{nullptr, 0};
+      bool usedForCount = false;
+
+      if (node->type == NODE_TYPE_REFERENCE) {
+        Variable const* v = static_cast<Variable const*>(node->getData());
+        auto setter = plan->getVarSetBy(v->id);
+        if (setter != nullptr && setter->getType() == EN::SUBQUERY) {
+          found.first = setter;
+          found.second = -1;
+        }
+      } else if (node->type == NODE_TYPE_INDEXED_ACCESS) {
+        auto sub = node->getMemberUnchecked(0);
+        if (sub->type == NODE_TYPE_REFERENCE) {
+          Variable const* v = static_cast<Variable const*>(sub->getData());
+          auto setter = plan->getVarSetBy(v->id);
+          auto index = node->getMemberUnchecked(1);
+          if (index->type == NODE_TYPE_VALUE && index->isNumericValue() &&
+              setter != nullptr && setter->getType() == EN::SUBQUERY) {
+            found.first = setter;
+            found.second = index->getIntValue() + 1;
+            if (found.second <= 0) {
+              found.second = -1;
+            }
+          }
+        }
+      } else if (node->type == NODE_TYPE_FCALL && node->numMembers() > 0) {
+        ast::FunctionCallNode fcall(node);
+        auto func = fcall.getFunction();
+        auto args = fcall.getArguments().getElements();
+        if (func->name == "FIRST" || func->name == "LENGTH" ||
+            func->name == "COUNT") {
+          if (args.size() > 0 && args[0]->type == NODE_TYPE_REFERENCE) {
+            ast::ReferenceNode ref(args[0]);
+            Variable const* v = ref.getVariable();
+            auto setter = plan->getVarSetBy(v->id);
+            if (setter != nullptr && setter->getType() == EN::SUBQUERY) {
+              found.first = setter;
+              if (func->name == "FIRST") {
+                found.second = 1;
+              } else {
+                found.second = -1;
+                usedForCount = true;
+              }
+            }
+          }
+        }
+      }
+
+      if (found.first != nullptr) {
+        auto it = subqueryAttributes.find(found.first);
+        if (it == subqueryAttributes.end()) {
+          subqueryAttributes.try_emplace(
+              found.first,
+              std::make_tuple(found.second,
+                              std::unordered_set<ExecutionNode const*>{n},
+                              usedForCount));
+        } else {
+          auto& sq = (*it).second;
+          if (usedForCount) {
+            std::get<2>(sq) = (std::get<0>(sq) <= 0);
+            std::get<0>(sq) = -1;
+            std::get<1>(sq).clear();
+          } else {
+            if (found.second <= 0 || std::get<0>(sq) < 0) {
+              std::get<0>(sq) = -1;
+              std::get<1>(sq).clear();
+            } else {
+              std::get<0>(sq) = std::max(std::get<0>(sq), found.second);
+              std::get<1>(sq).emplace(n);
+            }
+            std::get<2>(sq) = false;
+          }
+        }
+        return false;
+      }
+
+      return true;
+    };
+
+    Ast::traverseReadOnly(root, visitor, [](AstNode const*) {});
+  }
+
+  for (auto const& it : subqueryAttributes) {
+    ExecutionNode* node = it.first;
+    TRI_ASSERT(node->getType() == EN::SUBQUERY);
+    auto sn = ExecutionNode::castTo<SubqueryNode const*>(node);
+
+    if (sn->isModificationNode()) {
+      continue;
+    }
+
+    auto const& sq = it.second;
+    int64_t limitValue = std::get<0>(sq);
+    bool usedForCount = std::get<2>(sq);
+    if (limitValue <= 0 && !usedForCount) {
+      continue;
+    }
+
+    auto out = sn->outVariable();
+    VarSet used;
+    bool invalid = false;
+
+    auto current = node->getFirstParent();
+    while (current != nullptr) {
+      auto const& referencedBy = std::get<1>(sq);
+      if (referencedBy.find(current) == referencedBy.end()) {
+        used.clear();
+        current->getVariablesUsedHere(used);
+        if (used.find(out) != used.end()) {
+          invalid = true;
+          break;
+        }
+      }
+      current = current->getFirstParent();
+    }
+
+    if (invalid) {
+      continue;
+    }
+
+    auto root = sn->getSubquery();
+    if (root != nullptr && root->getType() == EN::RETURN) {
+      auto f = root->getFirstDependency();
+      TRI_ASSERT(f != nullptr);
+
+      if (std::get<2>(sq)) {
+        Ast* ast = plan->getAst();
+        auto expr =
+            std::make_unique<Expression>(ast, ast->createNodeValueBool(true));
+        Variable* outVariable = ast->variables()->createTemporaryVariable();
+        auto calcNode = plan->createNode<CalculationNode>(
+            plan.get(), plan->nextId(), std::move(expr), outVariable);
+        plan->insertAfter(f, calcNode);
+        TRI_ASSERT(root->getType() == EN::RETURN);
+        ExecutionNode::castTo<ReturnNode*>(root)->inVariable(outVariable);
+        modified = true;
+        continue;
+      }
+
+      if (f->getType() == EN::LIMIT) {
+        continue;
+      }
+
+      auto limitNode = plan->createNode<LimitNode>(plan.get(), plan->nextId(),
+                                                   0, limitValue);
+      plan->insertAfter(f, limitNode);
+      modified = true;
+    }
+  }
+
+  opt->addPlan(std::move(plan), rule, modified);
+}
+}  // namespace arangodb::aql
