+++ arangod/Aql/Optimizer/Rule/SpliceSubqueries.cpp
@@ -1,0 +1,237 @@
+////////////////////////////////////////////////////////////////////////////////
+/// DISCLAIMER
+///
+/// Copyright 2014-2026 ArangoDB GmbH, Cologne, Germany
+/// Copyright 2004-2014 triAGENS GmbH, Cologne, Germany
+///
+/// Licensed under the Business Source License 1.1 (the "License");
+/// you may not use this file except in compliance with the License.
+/// You may obtain a copy of the License at
+///
+///     https://github.com/arangodb/arangodb/blob/devel/LICENSE
+///
+/// Unless required by applicable law or agreed to in writing, software
+/// distributed under the License is distributed on an "AS IS" BASIS,
+/// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+/// See the License for the specific language governing permissions and
+/// limitations under the License.
+///
+/// Copyright holder is ArangoDB GmbH, Cologne, Germany
+///
+/// @author Max Neunhoeffer
+/// @author Jan Steemann
+////////////////////////////////////////////////////////////////////////////////
+
+#include "SpliceSubqueries.h"
+
+#include "Aql/ExecutionNode/ExecutionNode.h"
+#include "Aql/ExecutionNode/RemoteNode.h"
+#include "Aql/ExecutionNode/ReturnNode.h"
+#include "Aql/ExecutionNode/ScatterNode.h"
+#include "Aql/ExecutionNode/SubqueryEndExecutionNode.h"
+#include "Aql/ExecutionNode/SubqueryNode.h"
+#include "Aql/ExecutionNode/SubqueryStartExecutionNode.h"
+#include "Aql/ExecutionPlan.h"
+#include "Aql/Optimizer.h"
+#include "Aql/Query.h"
+#include "Aql/Variable.h"
+#include "Aql/Ast.h"
+#include "Aql/WalkerWorker.h"
+#include "Containers/SmallVector.h"
+#include "Containers/details/short_alloc.h"
+
+#include <set>
+
+namespace arangodb::aql {
+using EN = ExecutionNode;
+
+namespace {
+
+void findSubqueriesSuitableForSplicing(
+    ExecutionPlan const& plan,
+    containers::SmallVector<SubqueryNode*, 8>& result) {
+  TRI_ASSERT(result.empty());
+  using ResultVector = decltype(result);
+
+  using SuitableNodeSet =
+      std::set<SubqueryNode*, std::less<>,
+               containers::detail::short_alloc<SubqueryNode*, 128,
+                                               alignof(SubqueryNode*)>>;
+
+  class Finder final
+      : public WalkerWorker<ExecutionNode, WalkerUniqueness::NonUnique> {
+   public:
+    explicit Finder(ResultVector& result, SuitableNodeSet& suitableNodes)
+        : _result{result}, _suitableNodes{suitableNodes} {}
+
+    bool before(ExecutionNode* node) final {
+      TRI_ASSERT(node->getType() != EN::MUTEX);
+      if (node->getType() == ExecutionNode::SUBQUERY) {
+        _result.emplace_back(ExecutionNode::castTo<SubqueryNode*>(node));
+      }
+      return false;
+    }
+
+    bool enterSubquery(ExecutionNode*, ExecutionNode*) final {
+      ++_isSuitableLevel;
+      return true;
+    }
+
+    void leaveSubquery(ExecutionNode* subQuery, ExecutionNode*) final {
+      TRI_ASSERT(_isSuitableLevel != 0);
+      --_isSuitableLevel;
+      _suitableNodes.emplace(ExecutionNode::castTo<SubqueryNode*>(subQuery));
+    }
+
+   private:
+    ResultVector& _result;
+    SuitableNodeSet& _suitableNodes;
+    size_t _isSuitableLevel{1};
+  };
+
+  using SuitableNodeArena = SuitableNodeSet::allocator_type::arena_type;
+  SuitableNodeArena suitableNodeArena{};
+  SuitableNodeSet suitableNodes{suitableNodeArena};
+
+  auto finder = Finder{result, suitableNodes};
+  plan.root()->walkSubqueriesFirst(finder);
+
+  {
+    auto i = size_t{0};
+    auto j = size_t{0};
+    for (; j < result.size(); ++j) {
+      TRI_ASSERT(i <= j);
+      if (suitableNodes.count(result[j]) > 0) {
+        if (i != j) {
+          TRI_ASSERT(suitableNodes.count(result[i]) == 0);
+          result[i] = result[j];
+#ifdef ARANGODB_ENABLE_MAINTAINER_MODE
+          result[j] = nullptr;
+#endif
+        }
+        ++i;
+      }
+    }
+    TRI_ASSERT(i <= result.size());
+    result.resize(i);
+  }
+}
+
+}  // namespace
+
+void spliceSubqueriesRule(Optimizer* opt, std::unique_ptr<ExecutionPlan> plan,
+                          OptimizerRule const& rule) {
+  bool modified = false;
+
+  containers::SmallVector<SubqueryNode*, 8> subqueryNodes;
+  findSubqueriesSuitableForSplicing(*plan, subqueryNodes);
+
+  auto forAllDeps = [](ExecutionNode* node, auto cb) {
+    for (; node != nullptr; node = node->getFirstDependency()) {
+      TRI_ASSERT(node->getType() != ExecutionNode::SUBQUERY_START);
+      TRI_ASSERT(node->getType() != ExecutionNode::SUBQUERY_END);
+      cb(node);
+    }
+  };
+
+  for (auto const& sq : subqueryNodes) {
+    modified = true;
+    auto evenNumberOfRemotes = true;
+
+    forAllDeps(sq->getSubquery(), [&](auto node) {
+      node->setIsInSplicedSubquery(true);
+      if (node->getType() == ExecutionNode::REMOTE) {
+        evenNumberOfRemotes = !evenNumberOfRemotes;
+      }
+    });
+
+    auto const addClusterNodes = !evenNumberOfRemotes;
+
+    {
+      auto start = plan->createNode<SubqueryStartNode>(
+          plan.get(), plan->nextId(), sq->outVariable());
+
+      start->setIsInSplicedSubquery(sq->isInSplicedSubquery());
+
+      plan->insertBefore(sq, start);
+      TRI_ASSERT(start->getParents().size() == 1);
+      sq->removeDependency(start);
+      TRI_ASSERT(start->getParents().size() == 0);
+      TRI_ASSERT(start->getDependencies().size() == 1);
+      TRI_ASSERT(sq->getDependencies().size() == 0);
+      TRI_ASSERT(sq->getParents().size() == 1);
+
+      {
+        ExecutionNode* const singleton = sq->getSubquery()->getSingleton();
+        std::vector<ExecutionNode*> const parents = singleton->getParents();
+        TRI_ASSERT(parents.size() == 1);
+        auto oldSingletonParent = parents[0];
+        TRI_ASSERT(oldSingletonParent->getDependencies().size() == 1);
+        for (auto* x : parents) {
+          TRI_ASSERT(x != nullptr);
+          x->replaceDependency(singleton, start);
+        }
+        TRI_ASSERT(oldSingletonParent->getDependencies().size() == 1);
+        TRI_ASSERT(start->getParents().size() == 1);
+
+        if (addClusterNodes) {
+          auto const scatterNode = plan->createNode<ScatterNode>(
+              plan.get(), plan->nextId(), ScatterNode::SHARD);
+          auto const remoteNode = plan->createNode<RemoteNode>(
+              plan.get(), plan->nextId(), &plan->getAst()->query().vocbase(),
+              "", "", "");
+          scatterNode->setIsInSplicedSubquery(true);
+          remoteNode->setIsInSplicedSubquery(true);
+          plan->insertAfter(start, scatterNode);
+          plan->insertAfter(scatterNode, remoteNode);
+
+          TRI_ASSERT(remoteNode->getDependencies().size() == 1);
+          TRI_ASSERT(scatterNode->getDependencies().size() == 1);
+          TRI_ASSERT(remoteNode->getParents().size() == 1);
+          TRI_ASSERT(scatterNode->getParents().size() == 1);
+          TRI_ASSERT(oldSingletonParent->getFirstDependency() == remoteNode);
+          TRI_ASSERT(remoteNode->getFirstDependency() == scatterNode);
+          TRI_ASSERT(scatterNode->getFirstDependency() == start);
+          TRI_ASSERT(start->getFirstParent() == scatterNode);
+          TRI_ASSERT(scatterNode->getFirstParent() == remoteNode);
+          TRI_ASSERT(remoteNode->getFirstParent() == oldSingletonParent);
+        } else {
+          TRI_ASSERT(oldSingletonParent->getFirstDependency() == start);
+          TRI_ASSERT(start->getFirstParent() == oldSingletonParent);
+        }
+      }
+    }
+
+    {
+      ExecutionNode* subqueryRoot = sq->getSubquery();
+      Variable const* inVariable = nullptr;
+
+      if (subqueryRoot->getType() == ExecutionNode::RETURN) {
+        auto subqueryReturn = ExecutionNode::castTo<ReturnNode*>(subqueryRoot);
+        inVariable = subqueryReturn->inVariable();
+        TRI_ASSERT(subqueryReturn->getDependencies().size() == 1);
+        subqueryRoot = subqueryReturn->getFirstDependency();
+        TRI_ASSERT(!plan->isRoot(subqueryReturn));
+        plan->unlinkNode(subqueryReturn, true);
+      }
+
+      auto end = plan->createNode<SubqueryEndNode>(
+          plan.get(), plan->nextId(), inVariable, sq->outVariable());
+      end->setIsInSplicedSubquery(sq->isInSplicedSubquery());
+      plan->insertAfter(sq, end);
+
+      end->replaceDependency(sq, subqueryRoot);
+
+      TRI_ASSERT(end->getDependencies().size() == 1);
+      TRI_ASSERT(end->getParents().size() == 1);
+    }
+    TRI_ASSERT(sq->getDependencies().empty());
+    TRI_ASSERT(sq->getParents().empty());
+  }
+
+  if (modified) {
+    plan->root()->invalidateCost();
+  }
+  opt->addPlan(std::move(plan), rule, modified);
+}
+}  // namespace arangodb::aql
