+++ arangod/Aql/OptimizerRules.cpp
@@ -60,9 +60,7 @@
 #include "Aql/ExecutionNode/ScatterNode.h"
 #include "Aql/ExecutionNode/ShortestPathNode.h"
 #include "Aql/ExecutionNode/SortNode.h"
-#include "Aql/ExecutionNode/SubqueryEndExecutionNode.h"
 #include "Aql/ExecutionNode/SubqueryNode.h"
-#include "Aql/ExecutionNode/SubqueryStartExecutionNode.h"
 #include "Aql/ExecutionNode/TraversalNode.h"
 #include "Aql/ExecutionNode/MaterializeNode.h"
 #include "Aql/ExecutionNode/UpdateNode.h"
@@ -86,7 +84,6 @@
 #include "Aql/Variable.h"
 #include "Aql/types.h"
 #include "Basics/AttributeNameParser.h"
-#include "Basics/NumberUtils.h"
 #include "Basics/ScopeGuard.h"
 #include "Basics/StaticStrings.h"
 #include "Basics/SupervisedBuffer.h"
@@ -3685,249 +3682,6 @@ void arangodb::aql::optimizePathsRule(Optimizer* opt,
   opt->addPlan(std::move(plan), rule, modified);
 }
 
-/// @brief pulls out simple subqueries and merges them with the level above
-///
-/// For example, if we have the input query
-///
-/// FOR x IN (
-///     FOR y IN collection FILTER y.value >= 5 RETURN y.test
-///   )
-///   RETURN x.a
-///
-/// then this rule will transform it into:
-///
-/// FOR tmp IN collection
-///   FILTER tmp.value >= 5
-///   LET x = tmp.test
-///   RETURN x.a
-void arangodb::aql::inlineSubqueriesRule(Optimizer* opt,
-                                         std::unique_ptr<ExecutionPlan> plan,
-                                         OptimizerRule const& rule) {
-  containers::SmallVector<ExecutionNode*, 8> nodes;
-  plan->findNodesOfType(nodes, EN::SUBQUERY, true);
-
-  if (nodes.empty()) {
-    opt->addPlan(std::move(plan), rule, false);
-    return;
-  }
-
-  bool modified = false;
-  std::vector<ExecutionNode*> subNodes;
-
-  for (auto const& n : nodes) {
-    auto subqueryNode = ExecutionNode::castTo<SubqueryNode*>(n);
-
-    if (subqueryNode->isModificationNode()) {
-      // can't modify modifying subqueries
-      continue;
-    }
-
-    if (!subqueryNode->isDeterministic()) {
-      // can't inline non-deterministic subqueries
-      continue;
-    }
-
-    // check if subquery contains a COLLECT node with an INTO variable
-    // or a WINDOW node in an inner loop
-    bool eligible = true;
-    bool containsLimitOrSort = false;
-    auto current = subqueryNode->getSubquery();
-    TRI_ASSERT(current != nullptr);
-
-    while (current != nullptr) {
-      if (current->getType() == EN::WINDOW && subqueryNode->isInInnerLoop()) {
-        // WINDOW captures all existing rows in the scope, moving WINDOW
-        // ends up with different rows captured
-        eligible = false;
-        break;
-      } else if (current->getType() == EN::COLLECT) {
-        if (subqueryNode->isInInnerLoop()) {
-          eligible = false;
-          break;
-        }
-        if (ExecutionNode::castTo<CollectNode const*>(current)
-                ->hasOutVariable()) {
-          // COLLECT ... INTO captures all existing variables in the scope.
-          // if we move the subquery from one scope into another, we will end up
-          // with different variables captured, so we must not apply the
-          // optimization in this case.
-          eligible = false;
-          break;
-        }
-      } else if (current->getType() == EN::LIMIT ||
-                 current->getType() == EN::SORT) {
-        containsLimitOrSort = true;
-      }
-      current = current->getFirstDependency();
-    }
-
-    if (!eligible) {
-      continue;
-    }
-
-    Variable const* out = subqueryNode->outVariable();
-    TRI_ASSERT(out != nullptr);
-    // the subquery outvariable and all its aliases
-    VarSet subqueryVars;
-    subqueryVars.emplace(out);
-
-    // the potential calculation nodes that produce the aliases
-    std::vector<ExecutionNode*> aliasNodesToRemoveLater;
-
-    VarSet varsUsed;
-
-    current = n->getFirstParent();
-    // now check where the subquery is used
-    while (current->hasParent()) {
-      if (current->getType() == EN::ENUMERATE_LIST) {
-        if (current->isInInnerLoop() && containsLimitOrSort) {
-          // exit the loop
-          current = nullptr;
-          break;
-        }
-
-        // we're only interested in FOR loops...
-        auto listNode = ExecutionNode::castTo<EnumerateListNode*>(current);
-        if (listNode->getMode() == EnumerateListNode::kEnumerateObject) {
-          // exit the loop
-          current = nullptr;
-          break;
-        }
-        // ...that use our subquery as its input
-        if (subqueryVars.find(listNode->inVariable()) != subqueryVars.end()) {
-          // bingo!
-
-          // check if the subquery result variable or any of the aliases are
-          // used after the FOR loop
-          bool mustAbort = false;
-          for (auto const& itSub : subqueryVars) {
-            if (listNode->isVarUsedLater(itSub)) {
-              // exit the loop
-              current = nullptr;
-              mustAbort = true;
-              break;
-            }
-          }
-          if (mustAbort) {
-            break;
-          }
-
-          for (auto const& toRemove : aliasNodesToRemoveLater) {
-            plan->unlinkNode(toRemove, false);
-          }
-
-          subNodes.clear();
-          subNodes.reserve(4);
-          subqueryNode->getSubquery()->getDependencyChain(subNodes, true);
-          TRI_ASSERT(!subNodes.empty());
-          auto returnNode = ExecutionNode::castTo<ReturnNode*>(subNodes[0]);
-          TRI_ASSERT(returnNode->getType() == EN::RETURN);
-
-          modified = true;
-          auto queryVariables = plan->getAst()->variables();
-          auto previous = n->getFirstDependency();
-          auto insert = n->getFirstParent();
-          TRI_ASSERT(insert != nullptr);
-
-          // unlink the original SubqueryNode
-          plan->unlinkNode(n, false);
-
-          for (auto& it : subNodes) {
-            // first unlink them all
-            plan->unlinkNode(it, true);
-
-            if (it->getType() == EN::SINGLETON) {
-              // reached the singleton node already. that means we can stop
-              break;
-            }
-
-            // and now insert them one level up
-            if (it != returnNode) {
-              // we skip over the subquery's return node. we don't need it
-              // anymore
-              insert->removeDependencies();
-              TRI_ASSERT(it != nullptr);
-              insert->addDependency(it);
-              insert = it;
-
-              // additionally rename the variables from the subquery so they
-              // cannot conflict with the ones from the top query
-              for (auto const& variable : it->getVariablesSetHere()) {
-                queryVariables->renameVariable(variable->id);
-              }
-            }
-          }
-
-          // link the top node in the subquery with the original plan
-          if (previous != nullptr) {
-            insert->addDependency(previous);
-          }
-
-          // remove the list node from the plan
-          plan->unlinkNode(listNode, false);
-
-          queryVariables->renameVariable(returnNode->inVariable()->id,
-                                         listNode->outVariable()[0]->name);
-
-          // finally replace the variables
-          std::unordered_map<VariableId, Variable const*> replacements;
-          replacements.try_emplace(listNode->outVariable()[0]->id,
-                                   returnNode->inVariable());
-          VariableReplacer finder(replacements);
-          plan->root()->walk(finder);
-
-          plan->clearVarUsageComputed();
-          plan->findVarUsage();
-
-          // abort optimization
-          current = nullptr;
-        }
-      } else if (current->getType() == EN::CALCULATION) {
-        auto rootNode = ExecutionNode::castTo<CalculationNode*>(current)
-                            ->expression()
-                            ->node();
-        if (rootNode->type == NODE_TYPE_REFERENCE) {
-          if (subqueryVars.find(static_cast<Variable const*>(
-                  rootNode->getData())) != subqueryVars.end()) {
-            // found an alias for the subquery variable
-            subqueryVars.emplace(
-                ExecutionNode::castTo<CalculationNode*>(current)
-                    ->outVariable());
-            aliasNodesToRemoveLater.emplace_back(current);
-            current = current->getFirstParent();
-
-            continue;
-          }
-        }
-      }
-
-      if (current == nullptr) {
-        break;
-      }
-
-      varsUsed.clear();
-      current->getVariablesUsedHere(varsUsed);
-
-      bool mustAbort = false;
-      for (auto const& itSub : subqueryVars) {
-        if (varsUsed.find(itSub) != varsUsed.end()) {
-          // we found another node that uses the subquery variable
-          // we need to stop the optimization attempts here
-          mustAbort = true;
-          break;
-        }
-      }
-      if (mustAbort) {
-        break;
-      }
-
-      current = current->getFirstParent();
-    }
-  }
-
-  opt->addPlan(std::move(plan), rule, modified);
-}
-
 /// Essentially mirrors the geo::QueryParams struct, but with
 /// abstracts AstNode value objects
 struct GeoIndexInfo {
@@ -4676,203 +4430,6 @@ void arangodb::aql::geoIndexRule(Optimizer* opt,
   opt->addPlan(std::move(plan), rule, mod);
 }
 
-void arangodb::aql::optimizeSubqueriesRule(Optimizer* opt,
-                                           std::unique_ptr<ExecutionPlan> plan,
-                                           OptimizerRule const& rule) {
-  bool modified = false;
-
-  containers::SmallVector<ExecutionNode*, 8> nodes;
-  plan->findNodesOfType(nodes, EN::CALCULATION, true);
-
-  // value type is {limit value, referenced by, used for counting}
-  std::unordered_map<
-      ExecutionNode*,
-      std::tuple<int64_t, std::unordered_set<ExecutionNode const*>, bool>>
-      subqueryAttributes;
-
-  for (auto const& n : nodes) {
-    auto cn = ExecutionNode::castTo<CalculationNode*>(n);
-    AstNode const* root = cn->expression()->node();
-    if (root == nullptr) {
-      continue;
-    }
-
-    auto visitor = [&subqueryAttributes, &plan,
-                    n](AstNode const* node) -> bool {
-      std::pair<ExecutionNode*, int64_t> found{nullptr, 0};
-      bool usedForCount = false;
-
-      if (node->type == NODE_TYPE_REFERENCE) {
-        Variable const* v = static_cast<Variable const*>(node->getData());
-        auto setter = plan->getVarSetBy(v->id);
-        if (setter != nullptr && setter->getType() == EN::SUBQUERY) {
-          // we found a subquery result being used somehow in some
-          // way that will make the optimization produce wrong results
-          found.first = setter;
-          found.second = -1;  // negative values will disable the optimization
-        }
-      } else if (node->type == NODE_TYPE_INDEXED_ACCESS) {
-        auto sub = node->getMemberUnchecked(0);
-        if (sub->type == NODE_TYPE_REFERENCE) {
-          Variable const* v = static_cast<Variable const*>(sub->getData());
-          auto setter = plan->getVarSetBy(v->id);
-          auto index = node->getMemberUnchecked(1);
-          if (index->type == NODE_TYPE_VALUE && index->isNumericValue() &&
-              setter != nullptr && setter->getType() == EN::SUBQUERY) {
-            found.first = setter;
-            found.second = index->getIntValue() + 1;  // x[0] => LIMIT 1
-            if (found.second <= 0) {
-              // turn optimization off
-              found.second = -1;
-            }
-          }
-        }
-      } else if (node->type == NODE_TYPE_FCALL && node->numMembers() > 0) {
-        ast::FunctionCallNode fcall(node);
-        auto func = fcall.getFunction();
-        auto args = fcall.getArguments().getElements();
-        if (func->name == "FIRST" || func->name == "LENGTH" ||
-            func->name == "COUNT") {
-          if (args.size() > 0 && args[0]->type == NODE_TYPE_REFERENCE) {
-            ast::ReferenceNode ref(args[0]);
-            Variable const* v = ref.getVariable();
-            auto setter = plan->getVarSetBy(v->id);
-            if (setter != nullptr && setter->getType() == EN::SUBQUERY) {
-              found.first = setter;
-              if (func->name == "FIRST") {
-                found.second = 1;  // FIRST(x) => LIMIT 1
-              } else {
-                found.second = -1;
-                usedForCount = true;
-              }
-            }
-          }
-        }
-      }
-
-      if (found.first != nullptr) {
-        auto it = subqueryAttributes.find(found.first);
-        if (it == subqueryAttributes.end()) {
-          subqueryAttributes.try_emplace(
-              found.first,
-              std::make_tuple(found.second,
-                              std::unordered_set<ExecutionNode const*>{n},
-                              usedForCount));
-        } else {
-          auto& sq = (*it).second;
-          if (usedForCount) {
-            // COUNT + LIMIT together will turn off the optimization
-            std::get<2>(sq) = (std::get<0>(sq) <= 0);
-            std::get<0>(sq) = -1;
-            std::get<1>(sq).clear();
-          } else {
-            if (found.second <= 0 || std::get<0>(sq) < 0) {
-              // negative value will turn off the optimization
-              std::get<0>(sq) = -1;
-              std::get<1>(sq).clear();
-            } else {
-              // otherwise, use the maximum of the limits needed, and insert
-              // current node into our "safe" list
-              std::get<0>(sq) = std::max(std::get<0>(sq), found.second);
-              std::get<1>(sq).emplace(n);
-            }
-            std::get<2>(sq) = false;
-          }
-        }
-        // don't descend further
-        return false;
-      }
-
-      // descend further
-      return true;
-    };
-
-    Ast::traverseReadOnly(root, visitor, [](AstNode const*) {});
-  }
-
-  for (auto const& it : subqueryAttributes) {
-    ExecutionNode* node = it.first;
-    TRI_ASSERT(node->getType() == EN::SUBQUERY);
-    auto sn = ExecutionNode::castTo<SubqueryNode const*>(node);
-
-    if (sn->isModificationNode()) {
-      // cannot push a LIMIT into data-modification subqueries
-      continue;
-    }
-
-    auto const& sq = it.second;
-    int64_t limitValue = std::get<0>(sq);
-    bool usedForCount = std::get<2>(sq);
-    if (limitValue <= 0 && !usedForCount) {
-      // optimization turned off
-      continue;
-    }
-
-    // scan from the subquery node to the bottom of the ExecutionPlan to check
-    // if any of the following nodes also use the subquery result
-    auto out = sn->outVariable();
-    VarSet used;
-    bool invalid = false;
-
-    auto current = node->getFirstParent();
-    while (current != nullptr) {
-      auto const& referencedBy = std::get<1>(sq);
-      if (referencedBy.find(current) == referencedBy.end()) {
-        // node not found in "safe" list
-        // now check if it uses the subquery's out variable
-        used.clear();
-        current->getVariablesUsedHere(used);
-        if (used.find(out) != used.end()) {
-          invalid = true;
-          break;
-        }
-      }
-      // continue iteration
-      current = current->getFirstParent();
-    }
-
-    if (invalid) {
-      continue;
-    }
-
-    auto root = sn->getSubquery();
-    if (root != nullptr && root->getType() == EN::RETURN) {
-      // now inject a limit
-      auto f = root->getFirstDependency();
-      TRI_ASSERT(f != nullptr);
-
-      if (std::get<2>(sq)) {
-        Ast* ast = plan->getAst();
-        // generate a calculation node that only produces "true"
-        auto expr =
-            std::make_unique<Expression>(ast, ast->createNodeValueBool(true));
-        Variable* outVariable = ast->variables()->createTemporaryVariable();
-        auto calcNode = plan->createNode<CalculationNode>(
-            plan.get(), plan->nextId(), std::move(expr), outVariable);
-        plan->insertAfter(f, calcNode);
-        // change the result value of the existing Return node
-        TRI_ASSERT(root->getType() == EN::RETURN);
-        ExecutionNode::castTo<ReturnNode*>(root)->inVariable(outVariable);
-        modified = true;
-        continue;
-      }
-
-      if (f->getType() == EN::LIMIT) {
-        // subquery already has a LIMIT node at its end
-        // no need to do anything
-        continue;
-      }
-
-      auto limitNode = plan->createNode<LimitNode>(plan.get(), plan->nextId(),
-                                                   0, limitValue);
-      plan->insertAfter(f, limitNode);
-      modified = true;
-    }
-  }
-
-  opt->addPlan(std::move(plan), rule, modified);
-}
-
 /// @brief turn LENGTH(FOR doc IN ...) subqueries into an optimized count
 /// operation
 void arangodb::aql::optimizeCountRule(Optimizer* opt,
@@ -5278,238 +4835,6 @@ void arangodb::aql::activateCallstackSplit(ExecutionPlan& plan) {
 
   CallstackSplitter walker(options.maxNodesPerCallstack);
   plan.root()->walk(walker);
-}
-
-namespace {
-
-void findSubqueriesSuitableForSplicing(
-    ExecutionPlan const& plan,
-    containers::SmallVector<SubqueryNode*, 8>& result) {
-  TRI_ASSERT(result.empty());
-  using ResultVector = decltype(result);
-
-  using SuitableNodeSet =
-      std::set<SubqueryNode*, std::less<>,
-               containers::detail::short_alloc<SubqueryNode*, 128,
-                                               alignof(SubqueryNode*)>>;
-
-  // This finder adds all subquery nodes in pre-order to its `result` parameter,
-  // and all nodes that are suitable for splicing to `suitableNodes`. Suitable
-  // means that neither the containing subquery contains unsuitable nodes - at
-  // least not in an ancestor of the subquery - nor the subquery contains
-  // unsuitable nodes (directly, not recursively).
-  //
-  // It will be used in a fashion where the recursive walk on subqueries is done
-  // *before* the recursive walk on dependencies.
-  // It maintains a stack of bools for every subquery level. The topmost bool
-  // holds whether we've encountered a skipping block so far.
-  // When leaving a subquery, we decide whether it is suitable for splicing by
-  // inspecting the two topmost bools in the stack - the one belonging to the
-  // insides of the subquery, which we're going to pop right now, and the one
-  // belonging to the containing subquery.
-  //
-  // *All* subquery nodes will be added to &result in pre-order, and all
-  // *suitable* subquery nodes will be added to &suitableNodes. The latter can
-  // be omitted later, as soon as support for spliced subqueries / shadow rows
-  // is complete.
-
-  class Finder final
-      : public WalkerWorker<ExecutionNode, WalkerUniqueness::NonUnique> {
-   public:
-    explicit Finder(ResultVector& result, SuitableNodeSet& suitableNodes)
-        : _result{result}, _suitableNodes{suitableNodes} {}
-
-    bool before(ExecutionNode* node) final {
-      TRI_ASSERT(node->getType() != EN::MUTEX);  // should never appear here
-      if (node->getType() == ExecutionNode::SUBQUERY) {
-        _result.emplace_back(ExecutionNode::castTo<SubqueryNode*>(node));
-      }
-      return false;
-    }
-
-    bool enterSubquery(ExecutionNode*, ExecutionNode*) final {
-      ++_isSuitableLevel;
-      return true;
-    }
-
-    void leaveSubquery(ExecutionNode* subQuery, ExecutionNode*) final {
-      TRI_ASSERT(_isSuitableLevel != 0);
-      --_isSuitableLevel;
-      _suitableNodes.emplace(ExecutionNode::castTo<SubqueryNode*>(subQuery));
-    }
-
-   private:
-    // all subquery nodes will be added to _result in pre-order
-    ResultVector& _result;
-    // only suitable subquery nodes will be added to this set
-    SuitableNodeSet& _suitableNodes;
-    size_t _isSuitableLevel{1};  // push the top-level query
-  };
-
-  using SuitableNodeArena = SuitableNodeSet::allocator_type::arena_type;
-  SuitableNodeArena suitableNodeArena{};
-  SuitableNodeSet suitableNodes{suitableNodeArena};
-
-  auto finder = Finder{result, suitableNodes};
-  plan.root()->walkSubqueriesFirst(finder);
-
-  {  // remove unsuitable nodes from result
-    auto i = size_t{0};
-    auto j = size_t{0};
-    for (; j < result.size(); ++j) {
-      TRI_ASSERT(i <= j);
-      if (suitableNodes.count(result[j]) > 0) {
-        if (i != j) {
-          TRI_ASSERT(suitableNodes.count(result[i]) == 0);
-          result[i] = result[j];
-#ifdef ARANGODB_ENABLE_MAINTAINER_MODE
-          // To allow for the assert above
-          result[j] = nullptr;
-#endif
-        }
-        ++i;
-      }
-    }
-    TRI_ASSERT(i <= result.size());
-    result.resize(i);
-  }
-}
-}  // namespace
-
-// Splices in subqueries by replacing subquery nodes by
-// a SubqueryStartNode and a SubqueryEndNode with the subquery's nodes
-// in between.
-void arangodb::aql::spliceSubqueriesRule(Optimizer* opt,
-                                         std::unique_ptr<ExecutionPlan> plan,
-                                         OptimizerRule const& rule) {
-  bool modified = false;
-
-  containers::SmallVector<SubqueryNode*, 8> subqueryNodes;
-  findSubqueriesSuitableForSplicing(*plan, subqueryNodes);
-
-  // Note that we rely on the order `subqueryNodes` in the sense that, for
-  // nested subqueries, the outer subquery must come before the inner, so we
-  // don't iterate over spliced queries here.
-  auto forAllDeps = [](ExecutionNode* node, auto cb) {
-    for (; node != nullptr; node = node->getFirstDependency()) {
-      TRI_ASSERT(node->getType() != ExecutionNode::SUBQUERY_START);
-      TRI_ASSERT(node->getType() != ExecutionNode::SUBQUERY_END);
-      cb(node);
-    }
-  };
-
-  for (auto const& sq : subqueryNodes) {
-    modified = true;
-    auto evenNumberOfRemotes = true;
-
-    forAllDeps(sq->getSubquery(), [&](auto node) {
-      node->setIsInSplicedSubquery(true);
-      if (node->getType() == ExecutionNode::REMOTE) {
-        evenNumberOfRemotes = !evenNumberOfRemotes;
-      }
-    });
-
-    auto const addClusterNodes = !evenNumberOfRemotes;
-
-    {  // insert SubqueryStartNode
-
-      // Create new start node
-      auto start = plan->createNode<SubqueryStartNode>(
-          plan.get(), plan->nextId(), sq->outVariable());
-
-      // start and end inherit this property from the subquery node
-      start->setIsInSplicedSubquery(sq->isInSplicedSubquery());
-
-      // insert a SubqueryStartNode before the SubqueryNode
-      plan->insertBefore(sq, start);
-      // remove parent/dependency relation between sq and start
-      TRI_ASSERT(start->getParents().size() == 1);
-      sq->removeDependency(start);
-      TRI_ASSERT(start->getParents().size() == 0);
-      TRI_ASSERT(start->getDependencies().size() == 1);
-      TRI_ASSERT(sq->getDependencies().size() == 0);
-      TRI_ASSERT(sq->getParents().size() == 1);
-
-      {  // remove singleton
-        ExecutionNode* const singleton = sq->getSubquery()->getSingleton();
-        std::vector<ExecutionNode*> const parents = singleton->getParents();
-        TRI_ASSERT(parents.size() == 1);
-        auto oldSingletonParent = parents[0];
-        TRI_ASSERT(oldSingletonParent->getDependencies().size() == 1);
-        // All parents of the Singleton of the subquery become parents of the
-        // SubqueryStartNode. The singleton will be deleted after.
-        for (auto* x : parents) {
-          TRI_ASSERT(x != nullptr);
-          x->replaceDependency(singleton, start);
-        }
-        TRI_ASSERT(oldSingletonParent->getDependencies().size() == 1);
-        TRI_ASSERT(start->getParents().size() == 1);
-
-        if (addClusterNodes) {
-          auto const scatterNode = plan->createNode<ScatterNode>(
-              plan.get(), plan->nextId(), ScatterNode::SHARD);
-          auto const remoteNode = plan->createNode<RemoteNode>(
-              plan.get(), plan->nextId(), &plan->getAst()->query().vocbase(),
-              "", "", "");
-          scatterNode->setIsInSplicedSubquery(true);
-          remoteNode->setIsInSplicedSubquery(true);
-          plan->insertAfter(start, scatterNode);
-          plan->insertAfter(scatterNode, remoteNode);
-
-          TRI_ASSERT(remoteNode->getDependencies().size() == 1);
-          TRI_ASSERT(scatterNode->getDependencies().size() == 1);
-          TRI_ASSERT(remoteNode->getParents().size() == 1);
-          TRI_ASSERT(scatterNode->getParents().size() == 1);
-          TRI_ASSERT(oldSingletonParent->getFirstDependency() == remoteNode);
-          TRI_ASSERT(remoteNode->getFirstDependency() == scatterNode);
-          TRI_ASSERT(scatterNode->getFirstDependency() == start);
-          TRI_ASSERT(start->getFirstParent() == scatterNode);
-          TRI_ASSERT(scatterNode->getFirstParent() == remoteNode);
-          TRI_ASSERT(remoteNode->getFirstParent() == oldSingletonParent);
-        } else {
-          TRI_ASSERT(oldSingletonParent->getFirstDependency() == start);
-          TRI_ASSERT(start->getFirstParent() == oldSingletonParent);
-        }
-      }
-    }
-
-    {  // insert SubqueryEndNode
-
-      ExecutionNode* subqueryRoot = sq->getSubquery();
-      Variable const* inVariable = nullptr;
-
-      if (subqueryRoot->getType() == ExecutionNode::RETURN) {
-        // The SubqueryEndExecutor can read the input from the return Node.
-        auto subqueryReturn = ExecutionNode::castTo<ReturnNode*>(subqueryRoot);
-        inVariable = subqueryReturn->inVariable();
-        // Every return can only have a single dependency
-        TRI_ASSERT(subqueryReturn->getDependencies().size() == 1);
-        subqueryRoot = subqueryReturn->getFirstDependency();
-        TRI_ASSERT(!plan->isRoot(subqueryReturn));
-        plan->unlinkNode(subqueryReturn, true);
-      }
-
-      // Create new end node
-      auto end = plan->createNode<SubqueryEndNode>(
-          plan.get(), plan->nextId(), inVariable, sq->outVariable());
-      // start and end inherit this property from the subquery node
-      end->setIsInSplicedSubquery(sq->isInSplicedSubquery());
-      // insert a SubqueryEndNode after the SubqueryNode sq
-      plan->insertAfter(sq, end);
-
-      end->replaceDependency(sq, subqueryRoot);
-
-      TRI_ASSERT(end->getDependencies().size() == 1);
-      TRI_ASSERT(end->getParents().size() == 1);
-    }
-    TRI_ASSERT(sq->getDependencies().empty());
-    TRI_ASSERT(sq->getParents().empty());
-  }
-
-  if (modified) {
-    plan->root()->invalidateCost();
-  }
-  opt->addPlan(std::move(plan), rule, modified);
 }
 
 enum class DistributeType { DOCUMENT, TRAVERSAL, PATH };
