+++ arangod/Aql/OptimizerRulesFeature.cpp
@@ -48,6 +48,9 @@
 #include "Aql/Optimizer/Rule/OptimizeTraversals.h"
 #include "Aql/Optimizer/Rule/RemoveFiltersCoveredByTraversal.h"
 #include "Aql/Optimizer/Rule/RemoveTraversalPathVariable.h"
+#include "Aql/Optimizer/Rule/InlineSubqueries.h"
+#include "Aql/Optimizer/Rule/OptimizeSubqueries.h"
+#include "Aql/Optimizer/Rule/SpliceSubqueries.h"
 #include "Aql/OptimizerRules.h"
 #include "Aql/OptimizerRulesOptionsProvider.h"
 #include "Basics/Exceptions.h"
